+++ tests/Unit/Blocks/NumberedListItemTest.php
@@ -112,7 +112,7 @@ class NumberedListItemTest extends TestCase
         $expected = [
             "object"           => "block",
             "created_time"     => $i->block()->createdTime()->format(Date::FORMAT),
-            "last_edited_time" => $i->block()->lastEditedType()->format(Date::FORMAT),
+            "last_edited_time" => $i->block()->lastEditedTime()->format(Date::FORMAT),
             "archived"         => false,
             "has_children"     => false,
             "type"             => "numbered_list_item",