+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Fix typo ~~`Block::lastEditedType()`~~ `Block::lastEditedTime()`
+
 ## [0.6.0] - 2022-07-04
 
 ### Added